requirements.txt
# ---------------------------------------------------------------------------
# Project: voice2structured - Python依存関係
# Purpose: 必要なライブラリのバージョン管理
# ---------------------------------------------------------------------------

# 音声処理
pydub>=0.25.0

# Google Gemini API
google-generativeai>=0.8.0

# 設定ファイル処理
PyYAML>=6.0

# 非同期処理
asyncio-throttle>=1.0.2

# データ処理
pandas>=2.0.0

# ファイルシステム
pathlib2>=2.3.7